[sgen] Fix card scanning with larger block sizes.
The card scanning code assumed that the size of the block header is less
than the size of a card.  If we put the block info into the header and increase
the block size enough, that assumption isn't true anymore, because the
mark bits take up space in proportion to the block size.