[sgen] Always alloc degraded when nursery allocation fails.
There's an unlikely case that Roslyn triggers quite often on Linux: A
thread tries to allocate from the nursery, but it's full.  It triggers
a nursery collection, then tries allocating again.  Between the
collection and the allocation retry, another thread fills the nursery
so that the retry also fails.

We failed an assertion in this case.  Now we always go to the major
heap if it happens.