[jit] Use the proper function to get classes
For the opcodes sizeof, mkrefany and refanyval the jit
was resolving the operand tokens directly from the image.

In dynamic scenarios where the tokens are stored in the
wrapper data, this triggered TypeLoadExceptions.

This patch uses mini_get_class, as used everywhere else,
to resolve the operand tokens of those opcodes.