Unmanaged implementation of spin free locks.
The commit addresses race issues between the waiting on the lock semaphore and the release of the lock. The entry count and the owner field are grouped together into the status field. Accesses to it are done using atomic cas operations so threads aren't lost waiting for release on the lock semaphore.

Disable native fastpaths for now, until they are updated to the new implementation.