[msvc] Fixed heap problems at runtime when calling free().
Due to an incompatibility in the project files and DLL API setup code, we were importing locally defined symbols as DLL. This manifested as a debug heap dialog in VS caused by allocating and freeing memory between two Windows modules (which have their own memory heap).

Fix this by simplifying the DLL defines in MSVC, we are relying on .def files to manually specify the exports anyway.