Add DateTimeOffsetConverter to the default converters list.
This is the fix for https://bugzilla.xamarin.com/show_bug.cgi?id=25158

Current implementation does not contain the DateTimeOffsetConverter
in the list of default converters. This cause NotSupportedException
to be thrown when trying to serialize/deserialize DateTimeOffset
values to/from strings.

In .NET Framework implementation this class is added to such a list.