[sgen] Write barrier nursery checks might be needed
For the case of non-concurrent collector we don't want to mark stores from the heap to the heap, since it would add sgen overhead. On full aot we can't call the managed write barrier and we also don't now if the GC is concurrent or not, so we're being conservative.