HttpClientHandler shouldn't buffer requests
Turned off Allow(Read/Write)StreamBuffering to bring this inline with the .NET version of the HttpClientHandler which doesn't buffer request streams.  This took my team quite some time to track down where the problem was, when we were trying to upload larges files of 1GB or more.  We verified this fix with a custom build.