Fix command stream encoding for FTP requests.
Whenever a connection to an FTP server is established, FtpWebRequest
sends an option command to the server to enable UTF8 encoding for the
transfer. So far we've failed to check the request status of that
command and assumed that UTF8 is enabled but *also* failed to actually
use the encoding when sending commands (we had the ASCII encoding
hard-coded). This resulted in failed requests whenever a command named a
file or path which included non-ASCII characters. This patch does two
things:

 * checks the status of the UTF8 enabling command and uses UTF8 encoding
   only if it succeeded. Otherwise Encoding.Default is used.
 * uses the encoding detected above for the command stream.

Fixes https://bugzilla.xamarin.com/show_bug.cgi?id=24611

[Tests will follow at a later date]