Work in progress to bring Microsoft System.Runtime.Serialization to Mono
The current code compiles by replacing existing Mono code with
Microsoft code (without actually deleting the unreferenced files -
purely to use them as a roadmap for things that we should do to make
the code MOBILE and AOT friendly).

Currently this does not compile the mobile profile assembly, nor has
the code been tested yet.

The code also brings an extensive tracing/logging system that wont
work on Unix as it is based on Window's Event Tracing facility.

We should consider a few options:

   * Allow for the code to be disabled (for release builds on Mobile,
     or for an initial delivery that does not support any tracing
     capabilities).

   * Allow for the tracing framework to be replaced with some other
     logging framework.