[System.Net.Http] Fix CancelRequestViaProxy() test that intermittently fails on Jenkins
Seems that on Jenkins, the request to the (non-existing) IP address used in the test returns before the 1s timeout,
causing the test to fail because the cancellation didn't fire.

Lowering the timeout to 1 millisecond should fix this.

I also verified that the test still correctly fails when the commit that fixed the original issue is reverted.