Workaround the bug in the isinst opcode
This is another workaround the isinst opcode bug in Mono, similar to
the commit at acca62e1c3b5f1bff27f637e693f57a5417bf57d. In this case,
we need to make sure that we compare the unsigned array types first,
since the unsigned array types will match the signed array types when
the runtime is IL2CPP. The order of the checks will not have an impact
when the Mono runtime is used.