More changes to work around the isinst bug
With the il2cpp runtime (and its correct isinst implementation),
changing the order of these if checks as in my previous commit
(b4ddbbef1a3fb9338e375892258b9abbf780ad77) does correct the problem
with enums back by unsigned types. However, it introduced a new
problem. Now enums backed by signed types (notably, the default type,
Int32) incorrectly end up in the uint branch of this code. This causes
an InvalidCastException.

I’m not sure why the original unity-staging commit ABV did not catch
this, but the ABV with the unity-trunk commit did.

This change adds another check to each if statement, checking the type
of value as well. The isinst bug only occurs for array types, so the
check on value succeeds and allows the code to work in both cases.