[Sys.Data] fix DOM event order related issue caused by referencesource.
There is implementation mismatch in XmlDocument event order regarding
Load(). Take this code:

	string xml = "<root><child>1</child></root>";

	var doc = new XmlDocument ();
	doc.NodeInserted += (o, e) =>
		Console.WriteLine (e.Node.NodeType + e.Node.LocalName);
	doc.LoadXml (xml);

This is mono output before referencesource:

	Text#text
	Elementchild
	Elementroot

This is with referencesource:

	Elementchild
	Text#text
	Elementroot

So, there are event order differences in our two implementations.

Now, XmlDataDocument depends on those XmlDocument events to fill DataRow
values. It worked fine with pre-referencesource System.Xml model, but
with referencesource it skipped to fill data row values.

This change makes sure to fill data row values from text node which are
added _after_ the parent node is added to the document.