Work around the isinst problem in a different way
The last committed workaround of the isinst problem
(6fb387c638c055f5e5a886b3aa6549413926edc8) causing a consistent failure
of the
Unity.IntegrationTests.AssetImporting.TexturePostprocessing.TestPostproc
essingDefault integration test for the 64-bit editor on OSX. I cannot
reproduce this failure locally, and I cannot explain how this change
causes that problem.

So I have attempted to fix the problem in a different way, in hopes
that the integration test will pass. This is just a guess, but it is
the best option at the moment.