[Mono.Data.Tds] Deserialize variant-wrapped fixed-size integers (1-8 bytes)
Without this, variant output parameters are left unconsumed in the
input stream, and null is returned.

The new code skips the max length uint, and decodes a subset of the
variant types.  What is not recognized is skipped, and we return null
as the previous code used to do.

(Ideally, the subset of GetColumnValue which decodes types which can
be embedded in a variant should be factored out to a different method,
and reused from GetVariantValue.)