[System] Fix Interval_TooHigh_Enabled_Throw() test that failed on Linux
When the System.Timers.Timer code from referencesource was integrated, this test started failing on Linux.

It turns out there is no explicit check for Interval > Int32.MaxValue like we had in the Mono
implementation, instead it relies on the result of (int)Math.Ceiling(interval) in Timer.cs#L146
being negative and throws the exception inside the System.Threading.Timers code.

On Linux, the result of the call is 0 and thus no exception occurs.

Changing the test code to some other value greater than Int32.MaxValue fixes this (although it feels like a crutch).