[sgen] Fix clearing ephemerons and weak links on domain unload.
At some point we zeroed all the unloaded domain's objects first, and then we cleared
weak links, ephemerons, etc.  Because the objects were zeroes, we only had to
check that the vtable was null.  Now we do it the other way around: clear weak links,
ephemerons, ..., and then zero objects.  Obviously we now need to check for the
object's domain, because their vtables will not be null.