[metadata] Handle denormalized value type encoding of builtins.
All primitive types have a short encoding format that avoid using
a typeref token.

It's possible to encode them in another, longer, form with ELEMENT_TYPE_VALUETYPE + a typeref token.

Up to now, this was never seen in the wild so we never supported it.

This patch allows long form encoding of primitives.