[threading] Move jit_data from MonoThreadInternal to MonoThreadInfo. Cleanup jit attach code to support earlier attaching.
This move allows us to save the thread state much earlier on. In addition to the move, we now produce thread contexts without
the JIT TLS set. This might end up been an issue in some corner cases, just like with domain, so we'll address those when
they show up.