[WebClient/FtpWebRequest] Do not use MIME boundaries on FTP uploads (Fixes #26312)
The existing code only prevented HTTP PUTs from not having MIME-type
boundaries.   Apply this also to FTP methods.

Bring a referencesource helper method that determines the default
method based on a Uri.

The existing test case only tested for the attempt to upload a file
not for the contents.   Extended the code to check for a bit more
(just that the uplaod is actually 10 bytes long).

Fixes #26312