[Decimal] Completed the methods in decimal-ms.c
This completes the list of methods in decimal-ms.c that are surfaced
to the Decimal class.  This change also swaps out the Decimal class
from Mono to the Microsoft implementation.

This has not been tested yet, beyond the compiler being able to read
decimals (it did crash without it).

CoreCLR has two different bits of code for SearchScale, it looks like
one is just old and has not been updated.  Moved to the new one, and
updated the code to pass the extra data that is missing from the
former one.