initial attempt to import referencesource Encoding classes.
This does NOT work yet, because it's just to make everything to "build"
and there are missing InternalCall implementations (e.g. for
EncodingDataTable) and embedded resources (such as codepages.nlp).

I'm not sure what is the best solution for these classes yet. Creating
nlp resources for everything, or make some changes in Encoding and
import only partial sources such as UTF8Encoding.