[scripting] Make sure the exception object will not be collected prematurely (correct fix for case 661822)
This issue happens only in webplayer because only it relies on CoreCLR (the actual code path does not exist in other profiles) in which extra instructions are emitted for all method calls that do not have security attributes to ensure the code is allowed to call those methods. The extra instructions are just throwing an MethodAccessException but nothing is done to prevent the exception instance from being collected.