Add old mono Encoder and use it in MonoEncoding to keep compatibility.
There was some glitch in CP932 encoding and referencesource Encoder
that caused error in Convert(), shown in System.XML tests.
To avoid this, bring back old Encoder as we used before and keep
compatible processing.

It might be buggier, but doesn't regress. We may be able to bring nlp-
based Encodings later too.