Fix NTLM auth for requests with KeepAlive == false
NTLM Authentication requires connection to be maintained after successful
authentication. If client sets KeepAlive to false on an instance
of HttpWebRequest, when connecting to a server with NTLM authentication,
'Connection: close' header should NOT be included in the second step of handshake.

Before the fix, the fix, HttpWebRequest would keep
'Connection: Close' header, which would cause authentication failure.

More details in following article:
http://www.benjaminathawes.com/2010/10/14/ntlms-dependency-on-http-keep-alives-another-cause-of-the-dreaded-401-1-error/