Limit the lock to the writing operation
I initially discarded this change, because locking BaseSet (name, sec), didn't prevent the threads from trying to add the same key. That said, following the code, the point at which it actually fails is in NameObjectCollectionBase.BaseAdd, lines 331 and 332. The threads were arriving the second line at the same time:

   if (_entriesTable[name] == null)
          _entriesTable.Add(name, entry);

In brief, it's safe to call BaseSet twice with the same key, so it's safe to lock only the call to BaseSet