[SslClientStream] Add extra logging information, prevent crashes
Based on a number of exceptions being raised, we want to track down the reason
for some of the exception.   This patch both fixes the build (fixes the last
commit) and provides more information for the reason that the Alert is being
sent to the remote end.

The alert is currently being sent to the remote end when something fails in
the negotiation.   We likely need to *not send* alerts when the reason the
negotiation failed was that the connection was shutdown in some form (that is
why we have ObjectDisposed, or NotConnected exceptions).

This patch is a step towards verifying the theory.  This will help us
establish which step failed.   And the next step is determine why it failed.