[mdoc] Can now `mdoc update` multiple assemblies and assigne framework/platform ownership.
The biggest change is that the docs are now stored in the “unified” form on disk. After an `mdoc update`,
the XML files will be stored in folders that lack the “dropped namespace” (previously, it was stored in “classic” form, or with the dropped namespace). This makes it easier to have multiple frameworks. Speaking of which,
you can now pass in owner/platform information on the assembly list (see example below). This will mark
types and members with owner information that will be displayed in the rendered documentation.

Also, slightly changed the format for dropping namespaces on both the update and assemble
commands.

Owner … just add `=owner` to the assembly list:

    mdoc update Assembly1=Framework1 Assembly2=Framework2 -o en

Dropping namespaces … the assembly from which the namespace is being dropped is now added to the `dropns` parameter
along with the dropped namespace.:

    mdoc update Assembly1=Framework1 Assembly2=Framework1 --dropns=Assembly1=DroppedNS --dropns=Assembly2=OtherDroppedNS

Assembling classic doc sets after a namespace has been dropped. unified doesn’t need any special command line formulation, other than the fact that you no longer need to include the `dropns` parameter since it’s being handled in the classic update now:

    mdoc assemble en/ --droppedns=DroppedNS --droppedns=OtherDroppedNS --out=out-docs