Try to fix an assertion failure triggered by calls to System.Environment.Exit()
The comments and code for mono_walk_stack_full(...) seemed to be
inconsistent. The comments claimed that ``domain`` could be set to
NULL which would cause the current domain to be used. However the code
asserted that ``domain`` was never null which doesn't seem to be the
case.

Now if ``domain`` is set to NULL the value returned by mono_domain_get()
is used.

This seems to fix bug 28187