[mini] Fixed stack unwinding / DWARF register numbers when doing x86 cross compilation on OSX.
Eventually we want to add a better fix for this, like for instance adding this behaviour to the LLVM-specific code path
but until we can research more under exactly what conditions we need to do this then implement this simple workaround.

Fixes https://bugzilla.xamarin.com/show_bug.cgi?id=27055.